Move src/flat/* -> src/
