+++ src/exceptionobj.h
