+++ src/sharedarray.h
