+++ src/shareddict.h
