+++ src/sharedstr.h
